:sparkles: day 16 part 2

+++ src/main/java/krabban91/kodvent/kodvent/day16/OpCodes.java
@@ -2,35 +2,35 @@ package krabban91.kodvent.kodvent.day16;
 
 public interface OpCodes {
 
-    int[] addr(int[] register, int a, int b, int c);
+    long[] addr(long[] register, int a, int b, int c);
 
-    int[] addi(int[] register, int a, int b, int c);
+    long[] addi(long[] register, int a, int b, int c);
 
-    int[] mulr(int[] register, int a, int b, int c);
+    long[] mulr(long[] register, int a, int b, int c);
 
-    int[] muli(int[] register, int a, int b, int c);
+    long[] muli(long[] register, int a, int b, int c);
 
-    int[] banr(int[] register, int a, int b, int c);
+    long[] banr(long[] register, int a, int b, int c);
 
-    int[] bani(int[] register, int a, int b, int c);
+    long[] bani(long[] register, int a, int b, int c);
 
-    int[] borr(int[] register, int a, int b, int c);
+    long[] borr(long[] register, int a, int b, int c);
 
-    int[] bori(int[] register, int a, int b, int c);
+    long[] bori(long[] register, int a, int b, int c);
 
-    int[] setr(int[] register, int a, int b, int c);
+    long[] setr(long[] register, int a, int b, int c);
 
-    int[] seti(int[] register, int a, int b, int c);
+    long[] seti(long[] register, int a, int b, int c);
 
-    int[] gtir(int[] register, int a, int b, int c);
+    long[] gtir(long[] register, int a, int b, int c);
 
-    int[] gtri(int[] register, int a, int b, int c);
+    long[] gtri(long[] register, int a, int b, int c);
 
-    int[] gtrr(int[] register, int a, int b, int c);
+    long[] gtrr(long[] register, int a, int b, int c);
 
-    int[] eqir(int[] register, int a, int b, int c);
+    long[] eqir(long[] register, int a, int b, int c);
 
-    int[] eqri(int[] register, int a, int b, int c);
+    long[] eqri(long[] register, int a, int b, int c);
 
-    int[] eqrr(int[] register, int a, int b, int c);
+    long[] eqrr(long[] register, int a, int b, int c);
 }
